requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0

# LlamaIndex Framework
llama-index>=0.9.0
llama-index-core>=0.9.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0

# Hugging Face / Model Loading
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99

# Vector Database & Search
pymongo>=4.6.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Evaluation Metrics
ragas>=0.1.0
bert-score>=0.3.13
datasets>=2.16.0
nltk>=3.8.1
rouge-score>=0.1.2

# API Clients
requests>=2.31.0
httpx>=0.25.0
biopython>=1.83  # For PubMed E-utilities

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

# Demo & Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.8.0

# Utilities
pydantic>=2.5.0
rich>=13.7.0
tenacity>=8.2.3

# Development
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0
ipykernel>=6.27.0
jupyter>=1.0.0
